+++ system-parachains/encointer/src/lib.rs
@@ -220,7 +220,9 @@ impl InstanceFilter<RuntimeCall> for ProxyType {
 					RuntimeCall::EncointerReputationCommitments(_) |
 					RuntimeCall::EncointerFaucet(_) |
 					RuntimeCall::EncointerDemocracy(_) |
-					RuntimeCall::EncointerTreasuries(_)
+					RuntimeCall::EncointerTreasuries(_) |
+					RuntimeCall::EncointerOfflinePayment(_) |
+					RuntimeCall::EncointerReputationRings(_)
 			),
 			ProxyType::BazaarEdit => matches!(
 				c,
@@ -421,7 +423,6 @@ impl cumulus_pallet_parachain_system::Config for Runtime {
 	type CheckAssociatedRelayNumber = RelayNumberMonotonicallyIncreases;
 	type ConsensusHook = ConsensusHook;
 	type WeightInfo = weights::cumulus_pallet_parachain_system::WeightInfo<Runtime>;
-	type SelectCore = cumulus_pallet_parachain_system::DefaultCoreSelector<Runtime>;
 	type RelayParentOffset = ConstU32<0>;
 }
 
@@ -530,7 +531,8 @@ impl pallet_encointer_scheduler::Config for Runtime {
 	type RuntimeEvent = RuntimeEvent;
 	// attention!: EncointerDemocracy must be first hook as it potentially changes the rules for
 	// following hooks
-	type OnCeremonyPhaseChange = (EncointerDemocracy, EncointerCeremonies);
+	type OnCeremonyPhaseChange =
+		(EncointerDemocracy, EncointerCeremonies, EncointerReputationRings);
 	type MomentsPerDay = MomentsPerDay;
 	type CeremonyMaster = MoreThanHalfCouncil;
 	type WeightInfo = weights::pallet_encointer_scheduler::WeightInfo<Runtime>;
@@ -634,6 +636,32 @@ impl pallet_encointer_treasuries::Config for Runtime {
 	#[cfg(feature = "runtime-benchmarks")]
 	type BenchmarkHelper =
 		impls::benchmarks::TreasuryArguments<sp_core::ConstU8<1>, ConstU32<1000>>;
+}
+
+parameter_types! {
+	pub const MaxProofSize: u32 = 256;
+	pub const MaxVkSize: u32 = 2048;
+}
+
+impl pallet_encointer_offline_payment::Config for Runtime {
+	type RuntimeEvent = RuntimeEvent;
+	type WeightInfo = weights::pallet_encointer_offline_payment::WeightInfo<Runtime>;
+	type Currency = Balances;
+	type MaxProofSize = MaxProofSize;
+	type MaxVkSize = MaxVkSize;
+	type TrustedSetupOrigin = MoreThanHalfCouncil;
+}
+
+parameter_types! {
+	pub const MaxRingSize: u32 = 255;
+	pub const RingChunkSize: u32 = 100;
+}
+
+impl pallet_encointer_reputation_rings::Config for Runtime {
+	type RuntimeEvent = RuntimeEvent;
+	type WeightInfo = weights::pallet_encointer_reputation_rings::WeightInfo<Runtime>;
+	type MaxRingSize = MaxRingSize;
+	type ChunkSize = RingChunkSize;
 }
 
 impl pallet_aura::Config for Runtime {
@@ -848,6 +876,8 @@ construct_runtime! {
 		EncointerFaucet: pallet_encointer_faucet = 66,
 		EncointerDemocracy: pallet_encointer_democracy = 67,
 		EncointerTreasuries: pallet_encointer_treasuries = 68,
+		EncointerOfflinePayment: pallet_encointer_offline_payment = 69,
+		EncointerReputationRings: pallet_encointer_reputation_rings = 70,
 	}
 }
 
@@ -920,7 +950,7 @@ pub mod migrations {
 	use super::*;
 
 	/// Unreleased migrations. Add new ones here:
-	pub type Unreleased = ();
+	pub type Unreleased = (pallet_encointer_democracy::migrations::v2::MigrateV1toV2<Runtime>,);
 
 	/// All migrations that will run on the next runtime upgrade.
 	pub type SingleBlockMigrations = (Unreleased, Permanent);
@@ -965,7 +995,9 @@ mod benches {
 		[pallet_encointer_communities, EncointerCommunities]
 		[pallet_encointer_democracy, EncointerDemocracy]
 		[pallet_encointer_faucet, EncointerFaucet]
+		[pallet_encointer_offline_payment, EncointerOfflinePayment]
 		[pallet_encointer_reputation_commitments, EncointerReputationCommitments]
+		[pallet_encointer_reputation_rings, EncointerReputationRings]
 		[pallet_encointer_scheduler, EncointerScheduler]
 		[pallet_encointer_treasuries, EncointerTreasuries]
 		[cumulus_pallet_parachain_system, ParachainSystem]
@@ -1189,7 +1221,7 @@ impl_runtime_apis! {
 			VERSION
 		}
 
-		fn execute_block(block: Block) {
+		fn execute_block(block: <Block as BlockT>::LazyBlock) {
 			Executive::execute_block(block)
 		}
 
@@ -1225,7 +1257,7 @@ impl_runtime_apis! {
 		}
 
 		fn check_inherents(
-			block: Block,
+			block: <Block as BlockT>::LazyBlock,
 			data: sp_inherents::InherentData,
 		) -> sp_inherents::CheckInherentsResult {
 			data.check_extrinsics(&block)
@@ -1312,8 +1344,9 @@ impl_runtime_apis! {
 			PolkadotXcm::query_xcm_weight(message)
 		}
 
-		fn query_delivery_fees(destination: VersionedLocation, message: VersionedXcm<()>) -> Result<VersionedAssets, XcmPaymentApiError> {
-			PolkadotXcm::query_delivery_fees(destination, message)
+		fn query_delivery_fees(destination: VersionedLocation, message: VersionedXcm<()>, asset: VersionedAssetId) -> Result<VersionedAssets, XcmPaymentApiError> {
+			// We do not allow exchanging assets for delivery fees currently (our communities do not need to send XCMs paying with CC).
+			PolkadotXcm::query_delivery_fees::<()>(destination, message, asset)
 		}
 	}
 
@@ -1323,7 +1356,7 @@ impl_runtime_apis! {
 		}
 
 		fn dry_run_xcm(origin_location: VersionedLocation, xcm: VersionedXcm<RuntimeCall>) -> Result<XcmDryRunEffects<RuntimeEvent>, XcmDryRunApiError> {
-			PolkadotXcm::dry_run_xcm::<Runtime, xcm_config::XcmRouter, RuntimeCall, xcm_config::XcmConfig>(origin_location, xcm)
+			PolkadotXcm::dry_run_xcm::<xcm_config::XcmRouter>(origin_location, xcm)
 		}
 	}
 
@@ -1455,7 +1488,7 @@ impl_runtime_apis! {
 		}
 
 		fn execute_block(
-			block: Block,
+			block: <Block as BlockT>::LazyBlock,
 			state_root_check: bool,
 			signature_check: bool,
 			select: frame_try_runtime::TryStateSelect